# Improve Stream from Cats Effect Queue performance