@jordiolivares 2021-08-03
Seems like I didn't think of Scala 2.12 when using the Builder, let me quickly fix it